requirements.txt
# ML Pipeline Dependencies
google-generativeai==0.5.2
pdfplumber==0.11.0
pydantic==2.7.1

# Only needed if re-running data extraction
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
numpy==1.26.4
